fix cp i/o race condition
If we capped a point during overtime, we would still lose, because the I/O connection that adds time fires a frame late.  The master would see that all points are at full progress and make us lose.  We work around it by noting the last time a point was captured, and only determining points to be idle if we haven't capped in 100 ms and they are all at idle progress.