use new simulation time management
As noted in the commit in direct, the client's local frame time runs freely and is no longer affected by the simulation time.   We compute the delta between the server simulation time and the client frame time to convert between time contexts.  Since the client frame time and simulation time increase in lockstep, the delta should remain constant.